move standarize_entry to __init__.py
